fastapi_template/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
pydantic[email]==1.8.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
pytest==6.2.5
httpx==0.19.0
python-dotenv==0.19.0
asyncpg==0.24.0
alembic==1.7.3